+++ src/collision_object.cpp
@@ -12,7 +12,7 @@ std::vector<c_obj*> c_obj::objects;
 /**
  * @brief Construct a new c_obj object
  */
-c_obj::c_obj(sf::Texture& texture, sf::IntRect spriteTexture, sf::Vector2i pos) {
+c_obj::c_obj(sf::Texture& texture, sf::IntRect spriteTexture, sf::Vector2f pos) {
     mSprite.setTexture(texture);
     mTexture = &texture;
 
@@ -48,7 +48,7 @@ std::vector<c_obj*> c_obj::getCollided() {
  * 
  * @param pos Vector2i (x,y)
  */
-void c_obj::setPosition(sf::Vector2i pos) {
+void c_obj::setPosition(sf::Vector2f pos) {
     mSprite.setPosition(pos.x, pos.y);
 }
